more Mappings were added to UserController.java

+++ src/main/java/com/mealdelivery/food/structure/users/User.java
@@ -19,7 +19,7 @@ public class User {
     Long id;
     String name;
     String email;
-    Integer phone;
+    Long phone;
     Map<Integer, String> address;
     String hashedPassword;
     UserStatus userStatus;
